--- frontend/src/app/features/workers/update-workplace-details-after-staff-changes/update-staff-vacancies/update-staff-vacancies.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-update-staff-vacancies',
-  standalone: true,
-  imports: [],
-  templateUrl: './update-staff-vacancies.component.html',
-  styleUrl: './update-staff-vacancies.component.scss'
-})
-export class UpdateStaffVacanciesComponent {
-
-}
