+++ frontend/src/app/features/workers/update-workplace-details-after-staff-changes/update-vacancies-select-job-role/update-vacancies-select-job-role.component.html
