+++ frontend/src/app/features/workers/update-workplace-details-after-staff-changes/update-vacancies/update-vacancies.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-update-vacancies',
+  templateUrl: './update-vacancies.component.html',
+  styleUrl: './update-vacancies.component.scss',
+})
+export class UpdateVacanciesComponent {}
